exemplos/exemplo_CRUD_MQTT/requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-MQTT==1.2.1
PyMySQL==1.1.0
cryptography==41.0.7
Werkzeug==3.0.1
python-dotenv==1.0.0
